跳转任务消费模式

+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/SysAccount.java
@@ -76,11 +76,11 @@ public class SysAccount extends BaseEntity
     @Excel(name = "账号新增时间时区")
     private String sysAccountCreateTimezone;
 
-    private int sysAccountOnline;
+    private Long sysAccountOnline;
 
     private String sysAccountAbout;
 
-    private int sysAccountLoginStatus;
+    private Long sysAccountLoginStatus;
 
 
 }
